Renamed albums screen presentation entities, now they start with "feed" prefix. Moved getting albums tab content logic into single interactor
