--- sample/src/main/java/vsukharev/anytypeadapter/sample/albums/di/AlbumsScope.kt
@@ -1,7 +1,0 @@
-package vsukharev.anytypeadapter.sample.albums.di
-
-import javax.inject.Scope
-
-@Retention
-@Scope
-annotation class AlbumsScope
