--- sample/src/main/java/vsukharev/anytypeadapter/sample/albums/di/AlbumsScreenComponent.kt
@@ -1,11 +1,0 @@
-package vsukharev.anytypeadapter.sample.albums.di
-
-import dagger.Component
-import vsukharev.anytypeadapter.sample.albums.presentation.view.AlbumsFragment
-import vsukharev.anytypeadapter.sample.common.di.common.PerScreen
-
-@Component(dependencies = [AlbumsComponent::class])
-@PerScreen
-interface AlbumsScreenComponent {
-    fun inject(fragment: AlbumsFragment)
-}
