--- sample/src/main/java/vsukharev/anytypeadapter/sample/albums/presentation/view/AlbumsScreen.kt
@@ -1,10 +1,0 @@
-package vsukharev.anytypeadapter.sample.albums.presentation.view
-
-import androidx.fragment.app.Fragment
-import ru.terrakok.cicerone.android.support.SupportAppScreen
-
-class AlbumsScreen : SupportAppScreen() {
-    override fun getFragment(): Fragment {
-        return AlbumsFragment.newInstance()
-    }
-}
