+++ sample/src/main/java/vsukharev/anytypeadapter/sample/feed/data/FeedRepository.kt
@@ -1,0 +1,15 @@
+package vsukharev.anytypeadapter.sample.feed.data
+
+import vsukharev.anytypeadapter.sample.feed.di.FeedScope
+import vsukharev.anytypeadapter.sample.feed.domain.model.Feed
+import vsukharev.anytypeadapter.sample.common.network.makeSimpleRequest
+import javax.inject.Inject
+
+@FeedScope
+class FeedRepository @Inject constructor() {
+
+    suspend fun getFeed(): Feed {
+        makeSimpleRequest() // make request so that if there's no internet the error will be displayed
+        return Feed(AlbumsSource.albums, EditorsChoiceSource.editorsChoiceMusic)
+    }
+}
