+++ sample/src/main/java/vsukharev/anytypeadapter/sample/feed/di/FeedComponent.kt
@@ -1,0 +1,10 @@
+package vsukharev.anytypeadapter.sample.feed.di
+
+import dagger.Component
+import vsukharev.anytypeadapter.sample.feed.domain.interactor.FeedInteractor
+
+@Component
+@FeedScope
+interface FeedComponent {
+    fun albumsInteractor(): FeedInteractor
+}
