+++ sample/src/main/java/vsukharev/anytypeadapter/sample/feed/di/FeedScope.kt
@@ -1,0 +1,7 @@
+package vsukharev.anytypeadapter.sample.feed.di
+
+import javax.inject.Scope
+
+@Retention
+@Scope
+annotation class FeedScope
