+++ sample/src/main/java/vsukharev/anytypeadapter/sample/feed/domain/model/Album.kt
@@ -1,0 +1,12 @@
+package vsukharev.anytypeadapter.sample.feed.domain.model
+
+/**
+ * Model of a music album
+ */
+data class Album(
+    val id: String,
+    val name: String,
+    val performer: String,
+    val issueYear: Int,
+    val coverUrl: String
+)
