+++ sample/src/main/java/vsukharev/anytypeadapter/sample/feed/domain/model/EditorsChoice.kt
@@ -1,0 +1,13 @@
+package vsukharev.anytypeadapter.sample.feed.domain.model
+
+import java.util.*
+
+/**
+ * The model describing what the editor chooses
+ */
+data class EditorsChoice(
+    val id: UUID,
+    val starName: String,
+    val description: String,
+    val imageUrl: String
+)
