+++ sample/src/main/java/vsukharev/anytypeadapter/sample/feed/domain/model/Feed.kt
@@ -1,0 +1,6 @@
+package vsukharev.anytypeadapter.sample.feed.domain.model
+
+data class Feed(
+    val albums: List<Album>,
+    val editorsChoice: List<EditorsChoice>
+)
