+++ sample/src/main/res/layout/fragment_feed.xml
